Code-cancelNoReservationExists

+++ src/main/java/ReservationService.java
@@ -45,8 +45,12 @@ public class ReservationService {
      */
     public void cancel(String userId, String bookId) {
         // TODO: Implement using TDD
-        reservationRepo.delete(userId, bookId);
+        if (!reservationRepo.existsByUserAndBook(userId, bookId)) {
+            throw new IllegalArgumentException("No Reservation");
+        }
 
+        reservationRepo.delete(userId, bookId);
+        
         Book book = bookRepo.findById(bookId);
         book.setCopiesAvailable(book.getCopiesAvailable() + 1);
         bookRepo.save(book);
